move CarSearch into own file and component
